e2e: Proper relay cleanup + move seed fn to index
